backend/pyproject.toml
[project]
name = "timelineai-backend"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.14"
dependencies = [
    "aiosqlite>=0.22.1",
    "diskcache>=5.6.3",
    "fastapi>=0.128.6",
    "greenlet>=3.3.1",
    "httpx>=0.28.1",
    "ollama>=0.6.1",
    "piexif>=1.1.3",
    "pillow>=12.1.0",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "python-dateutil>=2.9.0.post0",
    "python-dotenv>=1.2.1",
    "sqlalchemy>=2.0.46",
    "uvicorn[standard]>=0.40.0",
]

[dependency-groups]
dev = [
    "black>=26.1.0",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "ruff>=0.15.0",
]
